backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.1.0",
    "fastapi>=0.115.12",
    "openai>=1.75.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
    "geopy>=2.4.1",
]
